Add files via upload
Here is a visual and practical demonstration of "Interrupts with LCD and STM32 MCU", a small project. This project displays the user's name initially, and after 5 seconds, it starts counting up every second. As soon as it receives the falling edge button interrupt, it turns on the on-board orange LED and notifies the  LCD by printing "Interrupt..."